requirements.txt
absl-py==0.12.0
altgraph==0.17.2
appdirs==1.4.4
argcomplete==1.12.3
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1618968359944/work
audioread==2.1.9
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
Bottleneck==1.3.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.3
cycler==0.10.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091796427/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1633213722787/work
filelock==3.0.12
gdown==3.13.1
gitdb==4.0.9
GitPython==3.1.26
h5py==3.4.0
htk-io==0.5
idna==3.2
interp-same==0.0.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1644979757530/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1645029586602/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1637175084646/work
joblib==1.0.1
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1642858610849/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1645024265313/work
kaldiio==2.17.2
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.37.0
matplotlib==3.4.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mcd==0.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
nara-wpe==0.0.7
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
noisereduce==2.0.0
numba==0.54.0
numpy==1.20.3
ortools==8.2.8710
packaging==21.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.3.2
pooch==1.5.1
praat-textgrids==1.3.1
progressbar==2.5
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1644497866770/work
protobuf==3.18.0
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pybullet==3.0.6
pycparser==2.20
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1641580240686/work
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
PyYAML==5.4.1
pyzmq==19.0.2
requests==2.26.0
resampy==0.2.2
scikit-learn==0.24.2
scipy==1.7.1
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
SoundFile==0.10.3.post1
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1644872665635/work
tensorboardX==2.4
threadpoolctl==2.2.0
toml==0.10.2
torch==1.9.1
torchaudio==0.7.0a0+ac17b64
torchcrepe==0.0.15
torchlibrosa==0.0.7
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094706440/work
tqdm==4.62.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
voicefixer==0.0.17
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
xmltodict==0.12.0
yq==2.12.2
